+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Send react native js engine, turbo module, fabric flags and component stack in Event contexts ([#2552](https://github.com/getsentry/sentry-react-native/pull/2552))
+
 ## 5.0.0-alpha.6
 
 - Latest changes from 4.6.1
 
 ### Features
+
 - Add initial support for the RN New Architecture, backwards compatible RNSentry Turbo Module ([#2522](https://github.com/getsentry/sentry-react-native/pull/2522))
 
 ### Breaking changes