+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Bump JavaScript SDK from v7.14.2 to v7.15.0 ([#2536](https://github.com/getsentry/sentry-react-native/pull/2536))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/master/CHANGELOG.md#7150)
   - [diff](https://github.com/getsentry/sentry-javascript/compare/7.14.2...7.15.0)
+- Bump Cocoa SDK from v7.27.1 to v7.28.0 ([#2548](https://github.com/getsentry/sentry-react-native/pull/2548))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/master/CHANGELOG.md#7280)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/7.27.1...7.28.0)
 
 ## 4.6.1